+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.application/pom.xml
@@ -1,51 +1,47 @@
 <?xml version="1.0" encoding="UTF-8"?>
-  <project 
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
-     http://maven.apache.org/xsd/maven-4.0.0.xsd"
-    xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-    
-  <modelVersion>4.0.0</modelVersion>
-  
-  <parent>
-    <groupId>com.abhsinh2.scpplugin</groupId>
-    <artifactId>com.abhsinh2.scpplugin.parent</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-  </parent>
-  
-  <groupId>com.abhsinh2.scpplugin</groupId>
-  <artifactId>com.abhsinh2.scpplugin.application</artifactId>
-  <packaging>eclipse-plugin</packaging>
-  
-  <build>
-  <plugins>
-    <plugin>
-      <groupId>org.eclipse.tycho</groupId>
-      <artifactId>tycho-p2-director-plugin</artifactId>   
-      <version>${tycho-version}</version>
-      <configuration>
-        <formats>
-          <win32>zip</win32>
-          <linux>tar.gz</linux>
-          <macosx>tar.gz</macosx>
-        </formats>
-      </configuration>
-      <executions>
-        <execution>
-          <id>materialize-products</id>
-          <goals>
-            <goal>materialize-products</goal>
-          </goals>
-        </execution>
-        <execution>
-          <id>archive-products</id>
-          <goals>
-            <goal>archive-products</goal>
-          </goals>
-        </execution>
-      </executions>
-    </plugin>
-  </plugins>
-</build>
-  
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>com.abhsinh2.scpplugin.application</artifactId>
+	<packaging>eclipse-plugin</packaging>
+
+	<parent>
+		<groupId>com.abhsinh2.scpplugin</groupId>
+		<artifactId>com.abhsinh2.scpplugin.parent</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
+
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>org.eclipse.tycho</groupId>
+				<artifactId>tycho-p2-director-plugin</artifactId>
+				<version>${tycho-version}</version>
+				<configuration>
+					<formats>
+						<win32>zip</win32>
+						<linux>tar.gz</linux>
+						<macosx>tar.gz</macosx>
+					</formats>
+				</configuration>
+				<executions>
+					<execution>
+						<id>materialize-products</id>
+						<goals>
+							<goal>materialize-products</goal>
+						</goals>
+					</execution>
+					<execution>
+						<id>archive-products</id>
+						<goals>
+							<goal>archive-products</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
 </project>
