+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.junit/pom.xml
@@ -1,109 +1,107 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/
-  POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" 
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  
-  	<groupId>com.abhsinh2.scpplugin</groupId>
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+	<modelVersion>4.0.0</modelVersion>
+
 	<artifactId>com.abhsinh2.scpplugin.junit</artifactId>
-	
-	<parent>
-  		<groupId>com.abhsinh2.scpplugin</groupId>
-  		<artifactId>com.abhsinh2.scpplugin.parent</artifactId>
-  		<version>1.0.0-SNAPSHOT</version>
-	</parent>
-	
 	<packaging>eclipse-test-plugin</packaging>
-	
+
+	<parent>
+		<groupId>com.abhsinh2.scpplugin</groupId>
+		<artifactId>com.abhsinh2.scpplugin.parent</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
+
 	<properties>
 		<tycho-version>0.18.0</tycho-version>
 	</properties>
-	
+
 	<repositories>
-  		<repository>
-    		<id>juno</id>
-    		<layout>p2</layout>
-    		<url>http://download.eclipse.org/releases/juno</url>
-  		</repository>
-  		
-  		<!-- repository>
-    		<id>kepler</id>
-    		<layout>p2</layout>
-    		<url>http://download.eclipse.org/releases/kepler</url>
-  		</repository -->
-  		
-  		<!-- repository>
-    		<id>luna</id>
-    		<layout>p2</layout>
-   			<url>http://download.eclipse.org/releases/luna</url>
-  		</repository -->
+		<repository>
+			<id>juno</id>
+			<layout>p2</layout>
+			<url>http://download.eclipse.org/releases/juno</url>
+		</repository>
+		
+		<!--
+		<repository> 
+			<id>kepler</id> 
+			<layout>p2</layout> 
+			<url>http://download.eclipse.org/releases/kepler</url> 
+		</repository>
+		-->
+	
+		<!--
+		<repository> 
+			<id>luna</id> 
+			<layout>p2</layout> 
+			<url>http://download.eclipse.org/releases/luna</url> 
+		</repository>
+		-->
 	</repositories>
-	
+
 	<profiles>
-	  <profile>
-	    <id>OSX</id>
-	    <activation>
-	      <os>
-	        <family>mac</family>
-	      </os>
-	    </activation>
-	    <properties>
-	      <swtbot.args>-Xmx1024m -XstartOnFirstThread</swtbot.args>
-	    </properties>
-	  </profile>
-	  <profile>
-	    <id>NotOSX</id>
-	    <activation>
-	      <os>
-	        <family>!mac</family>
-	      </os>
-	    </activation>
-	    <properties>
-	      <swtbot.args>-Xmx1024m</swtbot.args>
-	    </properties>
-	  </profile>
+		<profile>
+			<id>OSX</id>
+			<activation>
+				<os>
+					<family>mac</family>
+				</os>
+			</activation>
+			<properties>
+				<swtbot.args>-Xmx1024m -XstartOnFirstThread</swtbot.args>
+			</properties>
+		</profile>
+		<profile>
+			<id>NotOSX</id>
+			<activation>
+				<os>
+					<family>!mac</family>
+				</os>
+			</activation>
+			<properties>
+				<swtbot.args>-Xmx1024m</swtbot.args>
+			</properties>
+		</profile>
 	</profiles>
-	
+
 	<build>
 		<sourceDirectory>src</sourceDirectory>
-		
-	  <plugins>
-	    <plugin>
-	      <groupId>org.eclipse.tycho</groupId>
-	      <artifactId>tycho-maven-plugin</artifactId>
-	      <version>${tycho-version}</version>
-	      <extensions>true</extensions>
-	    </plugin>
-	    <plugin>
-	      <groupId>org.eclipse.tycho</groupId>
-	      <artifactId>tycho-surefire-plugin</artifactId>
-	      <version>${tycho-version}</version>
-	      <configuration>
-	        <useUIHarness>true</useUIHarness>
-	        <useUIThread>false</useUIThread>
-	        
-	        <dependencies>
-			    <dependency>
-			      <type>p2-installable-unit</type>
-			      <groupId>com.abhsinh2.scpplugin</groupId>
-			      <artifactId>com.abhsinh2.scpplugin.ui</artifactId>
-			    </dependency>
-			    <dependency>
-			      <type>p2-installable-unit</type>
-			      <artifactId>org.eclipse.sdk.feature.group</artifactId>
-			    </dependency>
-		  	</dependencies>
-		  	
-		  	<argLine>-XstartOnFirstThread</argLine>
-	        <product>org.eclipse.sdk.ide</product>
-	        <application>org.eclipse.ui.ide.workbench</application>
-	        
-	      </configuration>
-	    </plugin>
-	  </plugins>
-	  
-	  <pluginManagement>
+
+		<plugins>
+			<plugin>
+				<groupId>org.eclipse.tycho</groupId>
+				<artifactId>tycho-maven-plugin</artifactId>
+				<version>${tycho-version}</version>
+				<extensions>true</extensions>
+			</plugin>
+			<plugin>
+				<groupId>org.eclipse.tycho</groupId>
+				<artifactId>tycho-surefire-plugin</artifactId>
+				<version>${tycho-version}</version>
+				<configuration>
+					<useUIHarness>true</useUIHarness>
+					<useUIThread>false</useUIThread>
+					<dependencies>
+						<dependency>
+							<type>p2-installable-unit</type>
+							<groupId>com.abhsinh2.scpplugin</groupId>
+							<artifactId>com.abhsinh2.scpplugin.ui</artifactId>
+						</dependency>
+						<dependency>
+							<type>p2-installable-unit</type>
+							<artifactId>org.eclipse.sdk.feature.group</artifactId>
+						</dependency>
+					</dependencies>
+					<argLine>-XstartOnFirstThread</argLine>
+					<product>org.eclipse.sdk.ide</product>
+					<application>org.eclipse.ui.ide.workbench</application>
+				</configuration>
+			</plugin>
+		</plugins>
+
+		<pluginManagement>
 			<plugins>
 				<plugin>
 					<groupId>org.sonatype.tycho</groupId>
@@ -116,16 +114,16 @@
 				</plugin>
 				<plugin>
 					<groupId>org.sonatype.tycho</groupId>
-		    		<artifactId>tycho-versions-plugin</artifactId>
-      				<version>${tycho-version}</version>
-		      		<dependencies>
-          				<dependency>
-             				<groupId>org.eclipse.tigerstripe</groupId>
-             				<artifactId>tycho-versioning-extensions</artifactId>
-             				<version>0.0.1</version>
-          				</dependency>                           
-     				</dependencies>
- 				</plugin>				
+					<artifactId>tycho-versions-plugin</artifactId>
+					<version>${tycho-version}</version>
+					<dependencies>
+						<dependency>
+							<groupId>org.eclipse.tigerstripe</groupId>
+							<artifactId>tycho-versioning-extensions</artifactId>
+							<version>0.0.1</version>
+						</dependency>
+					</dependencies>
+				</plugin>
 				<plugin>
 					<artifactId>maven-compiler-plugin</artifactId>
 					<configuration>
@@ -133,10 +131,7 @@
 						<target>1.6</target>
 					</configuration>
 				</plugin>
-				
 			</plugins>
 		</pluginManagement>
 	</build>
-
-	
 </project>
