+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.ui/META-INF/MANIFEST.MF
@@ -3,7 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: SCP
 Bundle-SymbolicName: com.abhsinh2.scpplugin.ui;singleton:=true
 Bundle-Version: 1.0.0.qualifier
-Bundle-Activator: com.abhsinh2.plugin.scp.Activator
+Bundle-Activator: com.abhsinh2.scpplugin.ui.Activator
 Bundle-Vendor: ABHSINH2
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
