+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.ui/plugin.xml
@@ -41,6 +41,4 @@
             restorable="true">
       </view>
    </extension>
-
-
 </plugin>
