+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.ui/pom.xml
@@ -1,57 +1,57 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/
-  POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" 
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  
-  	<groupId>com.abhsinh2.scpplugin</groupId>
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+	<modelVersion>4.0.0</modelVersion>
+
 	<artifactId>com.abhsinh2.scpplugin.ui</artifactId>
-	<version>1.0.0-SNAPSHOT</version>
-	
 	<parent>
-  		<groupId>com.abhsinh2.scpplugin</groupId>
-  		<artifactId>com.abhsinh2.scpplugin.parent</artifactId>
-  		<version>1.0.0-SNAPSHOT</version>
+		<groupId>com.abhsinh2.scpplugin</groupId>
+		<artifactId>com.abhsinh2.scpplugin.parent</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
 	</parent>
-	
+
 	<packaging>eclipse-plugin</packaging>
-	
+
 	<properties>
 		<tycho-version>0.18.0</tycho-version>
 	</properties>
-	
+
 	<repositories>
-  		<repository>
-    		<id>juno</id>
-    		<layout>p2</layout>
-    		<url>http://download.eclipse.org/releases/juno</url>
-  		</repository>
-  		
-  		<!-- repository>
-    		<id>kepler</id>
-    		<layout>p2</layout>
-    		<url>http://download.eclipse.org/releases/kepler</url>
-  		</repository -->
-  		
-  		<!-- repository>
-    		<id>luna</id>
-    		<layout>p2</layout>
-   			<url>http://download.eclipse.org/releases/luna</url>
-  		</repository -->
+		<repository>
+			<id>juno</id>
+			<layout>p2</layout>
+			<url>http://download.eclipse.org/releases/juno</url>
+		</repository>
+		
+		<!--
+		<repository> 
+			<id>kepler</id> 
+			<layout>p2</layout> 
+			<url>http://download.eclipse.org/releases/kepler</url> 
+		</repository>
+		-->
+
+		<!--
+		<repository> 
+			<id>luna</id> 
+			<layout>p2</layout> 
+			<url>http://download.eclipse.org/releases/luna</url> 
+		</repository>
+		-->
 	</repositories>
-	
+
 	<build>
-	  <plugins>
-	    <plugin>
-	      <groupId>org.eclipse.tycho</groupId>
-	      <artifactId>tycho-maven-plugin</artifactId>
-	      <version>${tycho-version}</version>
-	      <extensions>true</extensions>
-	    </plugin>
-	  </plugins>
-	  
-	  <pluginManagement>
+		<plugins>
+			<plugin>
+				<groupId>org.eclipse.tycho</groupId>
+				<artifactId>tycho-maven-plugin</artifactId>
+				<version>${tycho-version}</version>
+				<extensions>true</extensions>
+			</plugin>
+		</plugins>
+
+		<pluginManagement>
 			<plugins>
 				<plugin>
 					<groupId>org.sonatype.tycho</groupId>
@@ -64,16 +64,16 @@
 				</plugin>
 				<plugin>
 					<groupId>org.sonatype.tycho</groupId>
-		    		<artifactId>tycho-versions-plugin</artifactId>
-      				<version>${tycho-version}</version>
-		      		<dependencies>
-          				<dependency>
-             				<groupId>org.eclipse.tigerstripe</groupId>
-             				<artifactId>tycho-versioning-extensions</artifactId>
-             				<version>0.0.1</version>
-          				</dependency>                           
-     				</dependencies>
- 				</plugin>				
+					<artifactId>tycho-versions-plugin</artifactId>
+					<version>${tycho-version}</version>
+					<dependencies>
+						<dependency>
+							<groupId>org.eclipse.tigerstripe</groupId>
+							<artifactId>tycho-versioning-extensions</artifactId>
+							<version>0.0.1</version>
+						</dependency>
+					</dependencies>
+				</plugin>
 				<plugin>
 					<artifactId>maven-compiler-plugin</artifactId>
 					<configuration>
@@ -84,6 +84,4 @@
 			</plugins>
 		</pluginManagement>
 	</build>
-
-	
 </project>
