--- com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.ui/src/com/abhsinh2/plugin/scp/model/SCPLocationManagerType.java
@@ -1,5 +1,0 @@
-package com.abhsinh2.plugin.scp.model;
-
-public enum SCPLocationManagerType {
-	ADDED, UPDATED, DELETED;
-}
