--- com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.ui/src/com/abhsinh2/plugin/scp/model/local/ISCPLocalLocation.java
@@ -1,15 +1,0 @@
-package com.abhsinh2.plugin.scp.model.local;
-
-import org.eclipse.core.runtime.IAdaptable;
-
-public interface ISCPLocalLocation extends IAdaptable {
-	public String getName();
-
-	public void setName(String newName);
-
-	public String getLocation();
-
-	public SCPLocalFileType getType();
-
-	public static ISCPLocalLocation[] NONE = new ISCPLocalLocation[] {};
-}
