+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.ui/src/com/abhsinh2/scpplugin/ui/model/SCPLocationManagerType.java
@@ -1,0 +1,5 @@
+package com.abhsinh2.scpplugin.ui.model;
+
+public enum SCPLocationManagerType {
+	ADDED, UPDATED, DELETED;
+}
