+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.ui/src/com/abhsinh2/scpplugin/ui/provider/SCPLocationTableLabelProvider.java
@@ -1,0 +1,28 @@
+package com.abhsinh2.scpplugin.ui.provider;
+
+import org.eclipse.jface.viewers.ITableLabelProvider;
+import org.eclipse.jface.viewers.LabelProvider;
+import org.eclipse.swt.graphics.Image;
+
+import com.abhsinh2.scpplugin.ui.model.remote.SCPLocation;
+
+public class SCPLocationTableLabelProvider extends LabelProvider implements
+		ITableLabelProvider {
+	public Image getColumnImage(Object element, int index) {
+		return null;
+	}
+
+	public String getColumnText(Object element, int index) {
+		SCPLocation location = (SCPLocation) element;
+		switch (index) {
+		case 0:
+			return location.getName();
+		case 1:
+			return location.getRemoteLocation().getRemoteAddress();
+		case 2:
+			return location.getRemoteLocation().getRemoteLocation();
+		default:
+			return "unknown " + index;
+		}
+	}
+}
