+++ com.abhsinh2.scpplugin.parent/com.abhsinh2.scpplugin.update/pom.xml
@@ -1,19 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/
-  POM/4.0.0 http://maven.apache.org/xsd/maven-
-  4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  
-  <modelVersion>4.0.0</modelVersion>
-  
-  <parent>
-    <groupId>com.abhsinh2.scpplugin</groupId>
-    <artifactId>com.abhsinh2.scpplugin.parent</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-  </parent>
-  
-  <groupId>com.abhsinh2.scpplugin</groupId>
-  <artifactId>com.abhsinh2.scpplugin.update</artifactId>
-  <packaging>eclipse-repository</packaging>
-  
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>com.abhsinh2.scpplugin.update</artifactId>
+	<packaging>eclipse-repository</packaging>
+
+	<parent>
+		<groupId>com.abhsinh2.scpplugin</groupId>
+		<artifactId>com.abhsinh2.scpplugin.parent</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
 </project>
